soc: qcom: ice: Prevent UFS probe deferral on ICE probe failure
When the ICE key programming interface is unavailable, the ice create
function fails, causing the probe to set NULL as the driver data. As a
result, when the UFS driver reads the ICE driver data and encounters a
NULL, leading to the deferral of the UFS probe and preventing the device
from booting to the shell.

To address this issue, modify the behavior to return an "operation not
supported" error when the ICE key programming interface is unavailable.
Additionally, mark this error in a global variable. When the UFS driver
attempts to read the ICE driver data, it will check for this error and
return it, rather than deferring the probe.

Signed-off-by: Yuvaraj Ranganathan <quic_yrangana@quicinc.com>